src/PGN/File/Syntax.php is created
