+++ tests/unit/PGN/File/data/01-non-str-games.pgn
