+++ tests/unit/PGN/File/data/01-text-with-non-str-games.pgn
