+++ tests/unit/PGN/File/data/02-non-str-games.pgn
