+++ tests/unit/PGN/File/data/02-text-with-non-str-games.pgn
