+++ tests/unit/PGN/File/data/03-non-str-games.pgn
