+++ tests/unit/PGN/File/data/03-text-with-non-str-games.pgn
